simplify to one regressors parameter
Because everything is now in one parameter, _make_denoiser and _set_denoiser have been removed. The logic inside _make_denoiser has been moved all into the set_regressors method. It's a lot simpler because there are now fewer combinations of inputs. `_load_from_strategy` handles the possibility that a strategy is not possible, and returns a dataframe to keep it consisent with how it was before (I think this way we're not left patching up weird downstream bugs in the report, etc).  There are also some minor stylistic changes I included that I didn't catch when I made the original code. Seems to work locally for me, but would be good to build up the test suite for this.